Added discrete notches to sliders, added slider tooltips

+++ Erovra3/src/engine/textureManager.c
@@ -18,9 +18,9 @@
 static Vector bezierCurve(float x[4], float y[4], double u)
 {
     float xu = 0.0f, yu = 0.0f;
-    xu = (float)(pow(1 - u, 3) * x[0] + 3 * u * pow(1 - u, 2) * x[1] + 3 * pow(u, 2) * (1 - u) * x[2]
+    xu = (int)(pow(1 - u, 3) * x[0] + 3 * u * pow(1 - u, 2) * x[1] + 3 * pow(u, 2) * (1 - u) * x[2]
         + pow(u, 3) * x[3]);
-    yu = (float)(pow(1 - u, 3) * y[0] + 3 * u * pow(1 - u, 2) * y[1] + 3 * pow(u, 2) * (1 - u) * y[2]
+    yu = (int)(pow(1 - u, 3) * y[0] + 3 * u * pow(1 - u, 2) * y[1] + 3 * pow(u, 2) * (1 - u) * y[2]
         + pow(u, 3) * y[3]);
     return (Vector) { xu, yu };
 }
